Reworking the render so that a menu item can always act as a root menu.
If you call render() on a menu item, it will be smart enough to render itself as the wrapping <ul> tag with its children inside of it.

This means that you can render the top-level menu item or any sub menu items and get the same results: a fully-formed unordered list.