Reworked depth not to use showChildren() but instead to be a non-mutating argument on ->render().
The ->render() method now takes an optional first argument, depth. A depth of 0 will return an empty string, a depth of 1 will return only the children. A depth of null will render to an infinite depth.

The showChildren() method can be used as before, but doesn't need to be depended on for limiting the rendering of menus only down to a certain level. The concern was that a user might hide the children lower than a certain level, render the top level menu, then try to render the sublevel menu somewhere else later. The sublevel menu would return a blank string because its children are still instructed to be hidden.

Now, showChildren() needs to only be used for discluding a particular leg of a menu.